Aggressive updates
Built the core of the combat loop to include the player_attack, and monster_attack functions.  This was a pretty aggressive piece of code and it has a few bugs.  I left some troubleshooting lines of code in the combat loop and the other functions in order to allow continued troubleshooting by any parties interested.